Change of type int to double

+++ Client/src/main/java/com/iv1201/client/model/ApplicationDTO.java
@@ -22,7 +22,7 @@ public class ApplicationDTO {
     
     @NotNull
     @Min(0)
-    private final int experience;
+    private final double experience;
     
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private final String start;
